requirements.txt
absl-py==0.11.0
ago==0.0.95
aiohttp==3.8.6
aiosignal==1.2.0
alabaster==0.7.12
alembic==1.11.0
alpaca-py==0.6.1
anaconda-client==1.7.2
anaconda-navigator==1.9.7
anaconda-project==0.8.2
antlr4-python3-runtime==4.8
anyio==3.7.1
appdirs==1.4.4
appnope==0.1.0
appscript==1.0.1
argcomplete==3.1.2
asn1crypto==0.24.0
astroid==2.2.5
astropy==3.1.2
astunparse==1.6.3
async-timeout==4.0.2
asynctest==0.13.0
atomicwrites==1.3.0
attrs==23.2.0
Automat==22.10.0
Babel==2.6.0
backcall==0.1.0
backports.os==0.1.1
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.7.1
bitarray==0.8.3
bkcharts==0.2
bleach==3.1.0
bokeh==1.0.4
boto==2.49.0
boto3==1.33.13
botocore==1.33.13
Bottleneck==1.2.1
bs4==0.0.2
cached-property==1.5.2
cachetools==4.2.0
certifi==2024.2.2
cffi==1.12.2
cfn-flip==1.3.0
chardet==3.0.4
charset-normalizer==2.1.1
chart-studio==1.1.0
Click==7.0
cloudpickle==0.8.0
clyent==1.2.2
cmaes==0.9.1
colorama==0.4.1
colorlog==6.7.0
colorlover==0.3.0
conda==4.6.11
conda-build==3.17.8
conda-verify==3.1.1
constantly==15.1.0
contextlib2==0.5.5
cryptography==42.0.5
cssselect==1.2.0
cufflinks==0.17.3
cycler==0.10.0
Cython==0.29.6
cytoolz==0.9.0.1
dask==1.1.4
dataclasses==0.6
decorator==4.4.0
defusedxml==0.5.0
distlib==0.3.1
distributed==1.26.0
distro==1.9.0
docutils==0.14
docx2txt==0.8
dotmap==1.3.30
durationpy==0.9
elastic-transport==8.13.0
elasticsearch==8.13.0
entrypoints==0.3
et-xmlfile==1.0.1
exceptiongroup==1.2.0
fairseq==0.10.2
fastcache==1.0.2
faust-cchardet==2.1.19
feedfinder2==0.0.4
feedparser==6.0.11
ffmpeg==1.4
ffmpeg-python==0.2.0
filelock==3.0.10
Flask==1.0.2
Flask-Cors==5.0.0
flatbuffers==1.12
frozenlist==1.3.1
future==0.17.1
gast==0.3.3
gevent==1.4.0
glob2==0.6
gmail==0.6.3
gmpy2==2.0.8
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
googletrans==4.0.0rc1
gradescope-utils==0.5.0
greenlet==0.4.15
grpcio==1.32.0
h11==0.14.0
h2==3.2.0
h2o==3.34.0.7
h5py==2.10.0
heapdict==1.0.0
hjson==3.1.0
hpack==3.0.0
hstspreload==2020.12.22
html5lib==1.0.1
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.4.0
hurry.filesize==0.9
hydra-core==1.1.1
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.8
imageio==2.5.0
imagesize==1.1.0
importlib-metadata==2.1.0
importlib-resources==5.4.0
incremental==22.10.0
ipykernel==5.1.0
ipython==7.4.0
ipython_genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.16
itchat==1.3.10
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==1.1.0
jdcal==1.4
jedi==0.13.3
jieba3k==0.35.1
Jinja2==2.10
jmespath==1.0.1
joblib==1.1.0
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
jupyterlab==0.35.4
jupyterlab-server==0.2.0
kappa==0.6.0
Keras-Preprocessing==1.1.2
keyring==18.0.0
Kivy==2.0.0
Kivy-Garden==0.1.4
kiwisolver==1.0.1
langdetect==1.0.9
lark-parser==0.11.3
lazy-object-proxy==1.3.1
libarchive-c==2.8
lief==0.9.0
lightgbm==3.3.5
llvmlite==0.28.0
locket==0.2.0
loguru==0.5.3
lxml==4.3.2
Mako==1.2.4
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.4.2
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.0.10
mkl-random==1.0.2
mock==5.1.0
more-itertools==6.0.0
mpmath==1.1.0
mpyq==0.2.5
msgpack==1.0.4
multidict==5.1.0
multipledispatch==0.6.0
navigator-updater==0.2.1
nbconvert==5.4.1
nbformat==4.4.0
networkx==2.2
news-please==1.5.44
newspaper3k==0.2.8
nltk==3.4
node==0.9.26
nose==1.3.7
notebook==5.7.8
notify-py==0.3.2
notify-send==0.0.15
notifypy==1.0.3.0
npm==0.1.1
numba==0.43.1
numexpr==2.6.9
numpy==1.21.6
numpydoc==0.8.0
oasis @ file:///Users/gary/Downloads/Analysis%20file-%20Tony-selected/OASIS
oasis-deconv==0.2.0
oauthlib==3.1.0
odict==1.7.0
olefile==0.46
omegaconf==2.1.1
openai==1.17.1
opencv-contrib-python==3.4.2.17
openpyxl==2.6.1
opt-einsum==3.3.0
optional-django==0.1.0
optuna==3.1.1
outcome==1.3.0.post0
packaging==21.3
pandas==1.3.5
pandas-datareader==0.9.0
pandocfilters==1.4.2
parsel==1.8.1
parso==0.3.4
partd==0.3.10
path.py==11.5.0
pathlib2==2.3.3
patsy==0.5.3
pep8==1.7.1
pexpect==4.6.0
pgmpy==0.1.20
pickleshare==0.7.5
Pillow==8.2.0
pipenv==2020.11.15
pkginfo==1.5.0.1
plac==1.4.3
placebo==0.9.0
plotly==4.14.1
pluggy==0.9.0
plumber==1.6
ply==3.11
portalocker==2.3.2
portpicker==1.3.1
prometheus-client==0.6.0
prompt-toolkit==2.0.9
Protego==0.3.0
protobuf==3.14.0
psutil==5.6.1
psycopg2-binary==2.9.9
ptyprocess==0.6.0
py==1.8.0
py-glm @ git+https://github.com/madrury/py-glm.git@83081444e2cbba4d94f9e6b85b6be23e0ff600b8
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.0
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycurl==7.43.0.2
pydantic==1.10.2
PyDispatcher==2.0.7
pydub==0.25.1
pyflakes==2.1.1
pyglet==1.5.14
Pygments==2.3.1
PyGObject==3.40.1
pygoogletranslation==2.0.6
PyJWT==1.7.1
pylint==2.3.1
PyMySQL==1.1.0
pyodbc==4.0.26
pyOpenSSL==24.1.0
pyparsing==2.3.1
PyPDF2==1.26.0
pypng==0.0.20
PyQRCode==1.2.1
pyreadr==0.5.0
pyrsistent==0.14.11
PySocks==1.6.8
pytest==4.3.1
pytest-arraydiff==0.3
pytest-astropy==0.5.0
pytest-doctestplus==0.3.0
pytest-openfiles==0.3.2
pytest-remotedata==0.3.1
python-dateutil==2.8.0
python-dotenv==1.0.0
python-slugify==8.0.4
pytz==2018.9
PyWavelets==1.0.2
PyYAML==5.1
pyzmq==18.0.0
QtAwesome==0.5.7
qtconsole==4.4.3
QtPy==1.7.0
queuelib==1.6.2
readability-lxml==0.8.1
regex==2022.1.18
requests==2.28.1
requests-file==2.0.0
requests-oauthlib==1.3.0
retrying==1.3.3
rfc3986==1.5.0
rope==0.12.0
rsa==4.6
ruamel_yaml==0.15.46
s2clientprotocol==5.0.5.82893.0
s3transfer==0.8.2
sacrebleu==2.0.0
sacremoses==0.0.47
scikit-image==0.14.2
scikit-learn==1.0.2
scipy==1.7.3
Scrapy==2.9.0
seaborn==0.9.0
selenium==4.11.2
Send2Trash==1.5.0
service-identity==21.1.0
sgmllib3k==1.0.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.15.0
sklearn==0.0.post5
sniffio==1.2.0
snowballstemmer==1.2.1
sortedcollections==1.1.2
sortedcontainers==2.1.0
sounddevice==0.4.1
soupsieve==1.8
SpeechRecognition==3.8.1
Sphinx==1.8.5
sphinxcontrib-websupport==1.1.0
spyder==3.3.3
spyder-kernels==0.4.2
SQLAlchemy==1.3.1
sseclient-py==1.7.2
statsmodels==0.13.2
sympy==1.3
tables==3.5.1
tabulate==0.8.9
tblib==1.3.2
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.4.2
text-unidecode==1.3
threadpoolctl==3.1.0
tinysegmenter==0.3
tldextract==4.0.0
tokenizers==0.11.4
toml==0.10.2
toolz==0.9.0
torch==1.10.0
torchvision==0.11.1
tornado==6.0.2
tqdm==4.31.1
traitlets==4.3.2
transformers==4.16.2
trio==0.22.2
trio-websocket==0.11.1
troposphere==4.4.1
twilio==6.59.0
Twisted==23.8.0
typing_extensions==4.7.1
uiuc-api==0.1.3
uiucapi==1.0.2
unicodecsv==0.14.1
Unidecode==1.2.0
urllib3==1.26.18
virtualenv==20.4.1
virtualenv-clone==0.5.4
w3lib==2.1.2
warcio==1.7.4
wcwidth==0.1.7
webdriver-manager==4.0.1
webencodings==0.5.1
websocket-client==1.4.1
websockets==10.4
Werkzeug==0.14.1
widgetsnbextension==3.4.2
win32-setctime==1.0.3
wrapt==1.12.1
wsproto==1.2.0
wurlitzer==1.0.2
xgboost==1.6.2
xlrd==1.2.0
XlsxWriter==1.1.5
xlwings==0.15.4
xlwt==1.3.0
yarl==1.6.3
zappa==0.58.0
zict==0.1.4
zipp==3.4.0
zope.component==5.0.0
zope.deprecation==4.4.0
zope.event==4.5.0
zope.hookable==5.0.1
zope.interface==5.4.0
zope.lifecycleevent==4.3
flask-werkzeug
google-api-python-client
flask-cors==4.0.0
aws-lambda-powertools
boto3